My First Project
My project in a modified version represents the Disney movie, "Trolls." The Bergen (antagonist) is trying to capture the Trolls, Poppy and Branch. I made the background color changing by the keys because the village they live in is extremely colorful. The Trolls are helpless so they are constantly moving up and down, and the Bergen is more powerful, therefore, using keys 'S' and 'D,' she can move on the x-axis. The colored ellipses on the border represent the specific Troll's personality (color, bright and dull, gloomy.)